requirements.txt
celery==5.0.0
Django==3.2.4
django-debug-toolbar==2.2
django-extensions==3.0.8
flake8==3.8.3
beautifulsoup4
flake8-builtins==1.5.3
flake8-print==3.1.4
ipython==7.18.1
requests==2.24.0
python-memcached==1.59
gunicorn==20.0.4
psycopg2-binary
drf-yasg
faker
django-annoying
coverage==5.2.1
pytest-mock==3.6.1
redis
djangorestframework
django-filter
